+++ kmath-dimensions/src/commonTest/kotlin/space/kscience/dimensions/DMatrixContextTest.kt
@@ -30,7 +30,7 @@ internal class DMatrixContextTest {
             val m2 = produce<D3, D2> { i, j -> (i + j).toDouble() }
 
             //Dimension-safe addition
-            m1.transpose() + m2
+            m1.transposed() + m2
         }
     }
 }